Added another unit test that pounds the SAME region with 3 different threads, rather than different regions.
This quickly showed a couple of synchronization problems in memory cache. I
changed some synchronization around and tried to to make the list remove cope
with the case where another thread has already removed the element. Before it
was just treating it like the only element, and quietly dropping the list (but
not the map -- yikes!)

Also, moved synchronization of spool to disk from Cache to LRUMemoryCache to
avoid a deadlock which occurred occasionally. Makes much more sense since that
method doesn't affect any of the volatile state in Cache.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/jcs/trunk@223922 13f79535-47bb-0310-9956-ffa450edef68