Modified and tested shrinking.
The concurrent modification exceptions were eating memory.

The shrinker was also removing items via the iterator, but the linked list still held references to them.

I reduced the chance of the exceptions, but there is still a race condition in the creation of the Array of keys.

The removals and gets go through the memory cache now.

We need to get rid of the MemoryCache interface and use the IMemoryCache interface.

There is a small memory usage by the shrinker, but the garbage collector is reclaiming it successfully.

I don't think there are any more memory leaks.  I'll test some more and then clean up.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/jcs/trunk@223936 13f79535-47bb-0310-9956-ffa450edef68