This had an awful bug.  All the events were getting sent into the Q and getting nullified.
The only reason the disk cache seemed to work was because of the purgatory.

This is one of the most central classes in the entire system.  It has been messed with a lot because it used to use JGL stuff.

It should be fixed now.

I tried to copy the Q to use for element events and found that the new implementation didn't work.

I guess this does everything the old Q used to do.  I'm not sure.

None of the auxiliaries could have worked properly since this was changed.

The take method was returning null always.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/jcs/trunk@223948 13f79535-47bb-0310-9956-ffa450edef68