Got rid of the JCSComponent based on stratum's component stuff, and removed the dependency on stratum.
Since turbine has decided to go with avalon instead of creating our own
component stuff, stratum will be disolved. I do not think it would be a
good idea to release with a dependency on it, hopefully an avalon component
can be in the next release.

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/jakarta/jcs/trunk@223998 13f79535-47bb-0310-9956-ffa450edef68