I saw the potential for a reported bug--where a get returns the wrong item under heavy load from a lateral request.  Although I do not recommend using laterals to get, it should work.
I put the operations that must occur in order in a synchronized block, insuring that gets from one lateral to another will not get mixed.  Needs to be improved, but should prevent any wrong elements from being returned.  . . .

Get timeout failures now zombie the lateral and start recovery.  Otherwise, since they do not occur in the background, the cache could grind to a halt.

The worst case now is that if another lateral gets hung, a caller trying to get will block until the timout.  Removes and puts are all synchronous calls.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/jcs/trunk@224076 13f79535-47bb-0310-9956-ffa450edef68