Fixed a bug that caused remote push to fail.
The event queues have an isAlive and an isWorking status.  Originally isAlive meant that it hadn't died.  No isAlive means are there any threads that are running.  isWorking used to mean that there were threads working; now it means that the queue is in working order.  This terminology isn't clear and lead to a bug.  I will come up with new names.

The remote server used to check to see if a queue was alive before trying to send an update to a local.  When we made the queues shutdown from inactivity, they started setting the isAlive to false.  The remote server would then think they were bad and not try to add events.  I now made the remote server check to see if the queue is working.

updated to bug fix version 1.2.1


git-svn-id: https://svn.apache.org/repos/asf/jakarta/jcs/trunk@224194 13f79535-47bb-0310-9956-ffa450edef68