Working on disk cache shutdown bug.  Items in memory were not making it to disk. The disk cache was killing the event queue and then optimizing the file and storing the keys.
I changed it so the disk cache will wait a configuratble number of seconds for the event queue to
empty out before shutting down.  (we still need a way to make the event queue stop accepting input during this process.)
There is a new disk cache config param
jcs.auxiliary.DC.attributes.ShutdownSpoolTimeLimit=60
The default is 60 seconds.  This is a per region setting.  It will wait this long for each region if it takes that long.

 . . .I want to add a setting for disk duplication of memory, so that all memory items will be backed on disk.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/jcs/trunk@224353 13f79535-47bb-0310-9956-ffa450edef68