Added system properties override functionality to the configure stage. You can set system properties that will be put into the cache properties prior to configuration. Added a simple test for the feature.
Removed soem unused interfaces.  There is inconsistency in the way interfaces are named.  . . .

Changed auxiliary cache interface to use the cache factory.  Many auxiliary listeners were
getting the factory by its get instance method.  This made it
almost impossible to unit test them.  I'm in the process of converting them all
to use the passed in cache manager.  This will allow us to mock
out the manager and the composite cache and create tests for the auxiliaries.
This is a work in progress.  The getCache method in the manager need to return an interface.  

git-svn-id: https://svn.apache.org/repos/asf/jakarta/jcs/trunk@230748 13f79535-47bb-0310-9956-ffa450edef68