Refactoring lateral manager and factory.  The old model was to have one factory and mananger know about all types of lateral.  This created too many dependencies and forced lots of ad hoc logic into the managers.  The new model will have each lateral responsible for its own factory and manager.  This keeps things clean and will allow laterals to ahve their own configuration settings.  The old attributes object had stuff for every lateral, but only a small subset applied to any given lateral.  This is a work in progress but it is almost done.
The new cache.ccf config for laterals will not be backward compatible, but the 
changes will be minor.  When it is done the version will be 1.3.  After more 
testing, this will be the release candidate major version.

I'm moving the javagroups stuff into to jdk1.4 specific jar.  

The refactoring should be done in a few days.  This will allow us to release a 1.3 core library
and a 1.4 extension jar.  

git-svn-id: https://svn.apache.org/repos/asf/jakarta/jcs/trunk@280254 13f79535-47bb-0310-9956-ffa450edef68