Added a unit test for a reported problem where an object with the same key but a different value is not sent across the tcp lateral cache properly.  I couldn't replicate the problem on jdk 1.3 or 1.4.
I suspect that the equals method of the user's object is problematic.  The object output stream tries to reuse
the serialed form of objects that it thinks are the same.  

If this turns out not to be the
case then we can either set the reset frequency to 1 or make it configurable.

With jdk1.4 there is a writeUnshared method that would prevent any such issues whatsoever.

git-svn-id: https://svn.apache.org/repos/asf/jakarta/jcs/trunk@332192 13f79535-47bb-0310-9956-ffa450edef68