I added a MySQL Disk Cache to JCS that extends the base JDBC Disk Cache. The MySQL child has two new features.  It can optimize on a schedule and balk.
The optimization schedule is a simple list of time in hh:mm:ss format. 
The optimizer knows how to detect optimization failures and will try to 
repair the table is the optimization fails.  
It also logs the table status after optimization and how long it took to perform the optimization.

It can also be configured to balk during optimizations.  
If you are running a JCS server cluster, it can be configured such that nodes will try to get from each other.  
If the mysql disk cache balks, the primary will to try and fetch data from the failover.  If the primary and failover are scheduled to optimize a few minute apart, then there should be no client impact.

I added a new pool access class and a table status object that is used to keep track of what processes might be working on a particular table.

I also changed the get to use toString rather than assume that the key is a string.

git-svn-id: https://svn.apache.org/repos/asf/jakarta/jcs/trunk@420956 13f79535-47bb-0310-9956-ffa450edef68