Fixed JCS-8 http://issues.apache.org/jira/browse/JCS-8
I registered the RemoteCacheManager as a shutdown observer with the CompositeCacheManager.  When the maanger shutdown method is called, it will release all the caches, un export the listiners, and tell the remote cache to deregister the listener.  

I also fixed the status indicator on the remote cache no wait facade.  This wasn't strictly necessary with the manager shutdown observer in place.

I also set the daemon status on the thread pool managed pools to true by default.

The remote cache client will now shutdown properly.

git-svn-id: https://svn.apache.org/repos/asf/jakarta/jcs/trunk@432936 13f79535-47bb-0310-9956-ffa450edef68