minor changes for testing
In order to make parts of the remote cache easier to test, I had to make some minor interface changes.  
I added a get attributes method to the auxiliary cache interface.  This made it necessary to change a few other files.
I added a new IRemoteCacheClient interface, so I could mock it out and test the remote cache no wait by itself.

To get started, I created a few new simple tests for the remote cache. 
The test coverage is still not very good, but it is up to 57% from 54%.

I also improved javadocs in several places. 

In addition, I removed an unused class in the remote package.

I started to take measures to standardize the interface naming convention.  
I don't like the "I prefix, but there are too many to make it worth chaning.  
In a subsequent release, we should deprecate the 3 or 4 interfaces that do not follow the I convention

git-svn-id: https://svn.apache.org/repos/asf/jakarta/jcs/trunk@436643 13f79535-47bb-0310-9956-ffa450edef68