Adding a rough version of a fixed block size disk cache.  Items are stored in blocks of a fixed size.  If they don't fit in a single block, they are put into multiple blocks.
It doesn't need optimization. The recycle bin is simply a list of integers that indicate what blocks are free.  The memory element just holds the key and the blocks.  

Compared to the Indexed Disk cache, this requires less memory overall, is easier to understand, and it does not require optimization.  

I'm still working on the BlockDiskCache, but it should be almost completely functional.



git-svn-id: https://svn.apache.org/repos/asf/jakarta/jcs/trunk@437310 13f79535-47bb-0310-9956-ffa450edef68