FIXED JCS-15
The indexed disk cache was not removing items when called with a partial key or a group. The problem was that it was trying to remove from an iterator on a copy of the keys.
I fixed the problem and added a couple unit tests.

I also cleaned up the block disk cache and change the lock.

git-svn-id: https://svn.apache.org/repos/asf/jakarta/jcs/trunk@519940 13f79535-47bb-0310-9956-ffa450edef68